separate `resource` from `types` (#372)
